requirements-mac.txt
pynput==1.7.3
pyautogui==0.9.53
opencv-python==4.5.5.64
mediapipe>=0.10.5
pyttsx3==2.71
SpeechRecognition==3.8.1
pyaudio==0.2.14
wikipedia==1.4.0
eel==0.14.0
googletrans==4.0.0-rc1
python-dotenv>=0.19.0
requests>=2.26.0
psutil>=5.8.0
# LangChain integration (intelligent fallback agent with memory)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
